Getting the data from the confirmation modal in Asp.Net Core part 159

+++ client/src/app/messages/messages.component.ts
@@ -3,6 +3,7 @@ import { Message } from '../_model/message';
 import { Pagination } from '../_model/pagination';
 import { MessageService } from '../_services/message.service';
 import { findIndex } from 'rxjs';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -19,6 +20,7 @@ export class MessagesComponent implements OnInit {
 
   constructor(
     private messageService: MessageService,
+    private confirmService: ConfirmService,
   ) { }
 
   ngOnInit(): void {
@@ -42,9 +44,16 @@ export class MessagesComponent implements OnInit {
       this.loadMessages();
     }
   }
+
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: _ => this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
-    })
+    this.confirmService.confirm('Confirm  delete message', 'This cannot be undone').subscribe({
+      next: result => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe({
+            next: _ => this.messages.splice(this.messages.findIndex(m => m.id === id), 1)
+          })
+        }
+      }
+    });
   }
 }
